Fix: Check Author before update"

+++ be-microservices/microservices/event-svc/src/modules/event/commands/DeleteEvent/deleteEvent.service.ts
@@ -8,8 +8,16 @@ export class DeleteEventService {
     private readonly deleteEventRepository: DeleteEventRepository,
   ) {}
 
-  async execute(id: number): Promise<Result<number, Error>> {
+  async execute(id: number, userId: string): Promise<Result<number, Error>> {
     try {
+      const isAuthor = await this.deleteEventRepository.checkAuthor(id, userId);
+      if (isAuthor.isErr()) {
+        return Err(new Error('Failed to check author'));
+      }
+      if (!isAuthor.unwrap()) {
+        return Err(new Error('Unauthorized'));
+      }
+
       const result = await this.deleteEventRepository.deleteEvent(id);
       if (result.isErr()) {
         return Err(new Error('Failed to delete event'));
