+++ be-microservices/microservices/event-svc/src/modules/showing/commands/deleteShowing/deleteShowing.service.ts
@@ -6,8 +6,16 @@ import { DeleteShowingRepository } from '../../repositories/deleteShowing.reposi
 export class DeleteShowingService {
   constructor(private readonly deleteShowingRepository: DeleteShowingRepository) {}
 
-  async execute(id: string): Promise<Result<string, Error>> {
+  async execute(id: string, userId: string): Promise<Result<string, Error>> {
     try {
+      const isAuthor = await this.deleteShowingRepository.checkAuthor(id, userId);
+      if (isAuthor.isErr()) {
+        return Err(new Error('Failed to check author'));
+      }
+      if (!isAuthor.unwrap()) {
+        return Err(new Error('Unauthorized'));
+      }
+
       const result = await this.deleteShowingRepository.deleteShowing(id);
       if (result.isErr()) {
         return Err(new Error('Failed to delete showing'));
