+++ be-microservices/microservices/event-svc/src/modules/ticketType/commands/deleteTicketType/deleteTicketType.service.ts
@@ -6,8 +6,15 @@ import { DeleteTicketTypeRepository } from '../../repositories/deleteTicketType.
 export class DeleteTicketTypeService {
   constructor(private readonly deleteTicketTypeRepository: DeleteTicketTypeRepository) {}
 
-  async execute(id: string): Promise<Result<string, Error>> {
+  async execute(id: string, userId: string): Promise<Result<string, Error>> {
     try {
+      const isAuthor = await this.deleteTicketTypeRepository.checkAuthor(id, userId);
+      if (isAuthor.isErr()) {
+        return Err(new Error('Failed to check author'));
+      }
+      if (!isAuthor.unwrap()) {
+        return Err(new Error('Unauthorized'));
+      }
       const result = await this.deleteTicketTypeRepository.deleteTicketType(id);
       if (result.isErr()) {
         return Err(new Error('Failed to delete ticket type'));
